delay the viee text dispatch until all text is collected
If a game has a type writer effect, VieeExtractor will send partial text that is gradually completed, we wait for a certain time before sending the pending text,This commit adds a mechanism to wait for a certain period before sending the pending text. We only dispatch the text to downstream consumers when no more text is received or when new text starts differently from the pending one.